refactor(Input): port to function component
The original class component used deprecated `componentWillReceiveProps` to change the
internal state of the component in response to an incoming prop change. Since the state
was merely mirroring the prop, I decided that there was no need to have internal state
at all anymore. In addition, the refactor to a hook, on top of removing internal state,
simplifies this component dramatically.