feat: add title page
Because the title and author comes from a different source than the transcripts,
we have to run the Fountain parser just before rendering the <Script /> component,
so that's where it lives now -- the fountain result is dispatched to store so that
it can be retrieved later (e.g. by an export) but content is not rendered from
there.