feat: basic mobile friendly layout, resolves #24
The difficulty in doing a truly mobile friendly layout is that the fountain.js script CSS which I've
borrowed only has two fixed sizes (well, technically, three, but not one small enough to fit on a mobile
phone). Inventing another fixed size would be catering to a lowest-common denominator where a script
page would only be, like, 320px wide, even on a phone that has space for a 500px wide element. We'd
want to eventually have a truly responsive page size, but then, the font would also need to adjust
to preserve letter-size page aspect ratio. Anyway: I'm kicking this can further down the road.

To implement basic mobile friendliness the two-column layout is now collapsed to one, below a 940px
breakpoint. (The CSS is also refactored to be mobile-first in its media queries.)

The script page preserves the 612px width (the smallest size already baked in) but we just allow the
phone to scroll horizontally to view this. While not the ideal end state, at least this is now
usuable and viewable on mobile, which is just where I want it for now.