feat: adjust styling for printing
Closes #22.

It's not perfect, but at least weird artifacts like the border around the script
won't be present during printing, and the UI will be hidden. Pagination and margins
will vary based on how the browser renders things for the printed page.

This required editing the original script.css styling so it's no longer an exact
copy of the original from Matt Daly's fountain.js demo page. This does simplify a few
other things -- the <Script /> component no longer needs to manually apply a class
style based on window width (we just let media queries take care of it) and the CSS
file will no longer be ignored by Prettier.